Améliorations diverses.

+++ diamond_square.pxd
@@ -1,4 +1,4 @@
 cimport numpy as np
 
 cpdef np.ndarray[double, ndim=2] build_height_map(
-        int size, int amplitude=*, int smoothing=*, bint save=*)
+    int size, int amplitude=*, float smoothing=*, bint save=*)
